move load/save to utils module
